# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Talk Proposals API",
        "description": "API documentation for the Talk Proposals system. This API allows speakers to submit talk proposals, reviewers to review them, and admins to manage the proposal status.",
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/reviews/rating-options": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get rating options",
                "description": "Returns all available rating options (1-5 and 10) with their labels.",
                "operationId": "1c1d3a7c76b2cc5e5b5c4b4c7e009c6f",
                "responses": {
                    "200": {
                        "description": "Rating options retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rating options retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "ratings": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "value": {
                                                                "type": "integer",
                                                                "example": 5
                                                            },
                                                            "label": {
                                                                "type": "string",
                                                                "example": "5 - Excellent"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/proposals/{proposalId}/reviews/{reviewId}": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get a specific review",
                "description": "Retrieves a single review by ID for a specific proposal.",
                "operationId": "e736ce7847eb5568fc122ca0cf57d81f",
                "parameters": [
                    {
                        "name": "proposalId",
                        "in": "path",
                        "description": "Proposal ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "reviewId",
                        "in": "path",
                        "description": "Review ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Review retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Review retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "review": {
                                                    "$ref": "#/components/schemas/Review"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Review or proposal not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Update a review",
                "description": "Updates an existing review. Only admins can update reviews. Rating must be 1, 2, 3, 4, 5, or 10.",
                "operationId": "225790f05749305524d735f034809c6a",
                "parameters": [
                    {
                        "name": "proposalId",
                        "in": "path",
                        "description": "Proposal ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "reviewId",
                        "in": "path",
                        "description": "Review ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rating"
                                ],
                                "properties": {
                                    "rating": {
                                        "description": "Rating value (1-5 or 10)",
                                        "type": "integer",
                                        "enum": [
                                            1,
                                            2,
                                            3,
                                            4,
                                            5,
                                            10
                                        ],
                                        "example": 5
                                    },
                                    "comment": {
                                        "description": "Optional review comment",
                                        "type": "string",
                                        "example": "Updated comment",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Review updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Review updated successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "review": {
                                                    "$ref": "#/components/schemas/Review"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized - Admin only"
                    },
                    "404": {
                        "description": "Review or proposal not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    },
                    "419": {
                        "description": "Missing or expired CSRF token for stateful SPA requests"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "admin",
                            "reviewer",
                            "speaker"
                        ],
                        "example": "speaker"
                    }
                },
                "type": "object"
            },
...
            "Review": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "rating": {
                        "type": "integer",
                        "enum": [
                            1,
                            2,
                            3,
                            4,
                            5,
                            10
                        ],
                        "example": 5
                    },
                    "comment": {
                        "type": "string",
                        "example": "Great proposal!",
                        "nullable": true
                    },
                    "reviewer": {
                        "$ref": "#/components/schemas/User"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Sanctum SPA session cookie. Fetch /sanctum/csrf-cookie before login and mutations, then send the X-XSRF-TOKEN header with cookies. The generated cookie name follows SESSION_COOKIE. Authenticate via the frontend or this API; session cookies cannot be entered manually in Swagger UI.",
                "name": "laravel-session",
                "in": "cookie"
            },
            "bearerAuth": {
                "type": "http",
                "description": "An existing Sanctum personal access token. Login/register use SPA sessions and do not issue bearer tokens.",
                "scheme": "bearer"
            }
        }
    },
...
}